+++ src/main/java/net/kaczmarzyk/spring/data/jpa/web/SpecificationArgumentResolver.java
@@ -16,6 +16,7 @@
 package net.kaczmarzyk.spring.data.jpa.web;
 
 import net.kaczmarzyk.spring.data.jpa.utils.TypeUtil;
+import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.core.MethodParameter;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.data.jpa.domain.Specification;
@@ -42,11 +43,19 @@ public class SpecificationArgumentResolver implements HandlerMethodArgumentResol
 	private static Map<Class<? extends Annotation>, SpecificationResolver<? extends Annotation>> resolversBySupportedType;
 
 	public SpecificationArgumentResolver() {
-		 this(null);
+		 this(null, null);
 	}
 	
 	public SpecificationArgumentResolver(ConversionService conversionService) {
-		SimpleSpecificationResolver simpleSpecificationResolver = new SimpleSpecificationResolver(conversionService);
+		this(conversionService, null);
+	}
+	
+	public SpecificationArgumentResolver(AbstractApplicationContext applicationContext) {
+		this(null, applicationContext);
+	}
+	
+	public SpecificationArgumentResolver(ConversionService conversionService, AbstractApplicationContext abstractApplicationContext) {
+		SimpleSpecificationResolver simpleSpecificationResolver = new SimpleSpecificationResolver(conversionService, abstractApplicationContext);
 		
 		resolversBySupportedType = Arrays.asList(
 				simpleSpecificationResolver,