+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+v2.6.0 UNRELEASED
+======
+* Added support for [SpEL](https://docs.spring.io/spring/docs/5.2.7.RELEASE/spring-framework-reference/core.html#expressions) and [property placeholders](https://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/context/support/PropertySourcesPlaceholderConfigurer.html) in `@Spec` attributes: `constVal`, `defaultVal`. 
+  To enable SpEL support in particular `@Spec` attribute:
+  * Configure `SpecificationArgumentResolver` by passing `AbstractApplicationContext` in constructor
+  * Set attribute `constValType`/`defaultValType` value to `SpEL`
+  
+  Configuration example:
+  ```java
+  	@Autowired
+  	AbstractApplicationContext applicationContext;
+  	
+  	@Override
+  	public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+  		argumentResolvers.add(new SpecificationArgumentResolver(applicationContext));
+  	}
+  ```
+
+  Example of default value with property placeholder  #1:
+
+  ```java
+  @RequestMapping(value = "/customers")
+  @ResponseBody
+  public Object findByLastName(
+          @Spec(path = "id", params="lastName", defaultVal='${search.default-params.lastName}', defaultValType='SpEL', spec = Equal.class) Specification<Customer> spec) {
+  	
+  	return customerRepo.findAll(spec);
+  }
+  ```
+
+  application.properties
+  ```properties
+  search.default-params.lastName=Simpson
+  ```
+
+  Example of default value in [SpEL](https://docs.spring.io/spring/docs/5.2.7.RELEASE/spring-framework-reference/core.html#expressions) #2:
+
+  ```java
+  @RequestMapping(value = "/customers")
+  @ResponseBody
+  public Object findCustomersWhoCameFromTheFuture(
+          @Spec(path = "id", params="birthDate", defaultVal='#{T(java.time.LocalDate).now()}', defaultValType='SpEL', spec = GreaterThanOrEqual.class) Specification<Customer> spec) {
+  	
+  	return customerRepo.findAll(spec);
+  }
+  ```
+
 v2.5.0
 ======
 * Added support for repeatable `@Join` and `@JoinFetch` annotations. `@Joins` annotation is now deprecated and it's going to be removed in the future.